Adding a 2nd property on the Siena benchmark
