--- java/java-ranger-regression/siena_eqchk.yml
@@ -1,8 +1,0 @@
-format_version: "1.0"
-input_files:
-  - ../common/
-  - siena_eqchk/
-properties:
-  - property_file: ../properties/assert.prp
-    expected_verdict: true
-
