+++ java/java-ranger-regression/siena_eqchk/siena_prop2.yml
@@ -1,0 +1,9 @@
+format_version: "1.0"
+input_files:
+  - ../../common/
+  - impl/
+  - prop2/
+properties:
+  - property_file: ../../properties/assert.prp
+    expected_verdict: false
+
